cicd update

+++ .circleci/doTests.sh
@@ -145,6 +145,11 @@ while [ $i -lt $frontendDriverLength ]; do
     ./setupAndTestWithFrontend.sh $coreCommercial $frontendTag $nodeTag
 
     someFrontendTestsRan=true
+    if [[ $? -ne 0 ]]
+    then
+        echo "test failed... exiting!"
+        exit 1
+    fi
     rm -rf ../../com-root
 done
